+++ src/main/kotlin/kr/co/amateurs/mcp/server/post/dto/request/PostSearchParams.kt
@@ -1,10 +1,12 @@
 package kr.co.amateurs.mcp.server.post.dto.request
 
+import jakarta.validation.constraints.NotBlank
 import kr.co.amateurs.mcp.server.common.dto.PaginationParam
 import kr.co.amateurs.mcp.server.common.dto.PaginationSortType
 import org.springframework.data.domain.Sort
 
 class PostSearchParams(
+    @field:NotBlank("키워드 혹은 닉네임 없이 비어있는 값을 입력할 수 없습니다")
     val keyword: String,
     override val pageNumber: Int = 0,
     override val pageSize: Int = 10,
