+++ src/main/kotlin/kr/co/amateurs/mcp/server/post/repository/query/PostQuery.kt
@@ -1,0 +1,5 @@
+package kr.co.amateurs.mcp.server.post.repository.query
+
+interface PostQuery<T> {
+    fun execute(): T
+}
