+++ src/main/kotlin/kr/co/amateurs/mcp/server/post/service/PostService.kt
@@ -12,22 +12,29 @@ class PostService(
 ) {
     @Tool(
         name = "search_posts", description = """
-        
+        키워드와 일치하는 게시글 목록을 검색합니다.
     """
     )
     fun searchPosts(
-        @ToolParam
-        postSearchParams: PostSearchParams
-    ) = postRepository.findAll(postSearchParams)
+        @ToolParam(
+            description = """
+                keyword: 검색할 키워드
+            """
+        )
+        params: PostSearchParams
+    ) = postRepository.findAll(params)
 
     @Tool(
         name = "search_posts_by_nickname", description = """
-        
+        유저의 닉네임과 일치하는 게시글 목록을 검색합니다.
     """
     )
     fun searchByNickname(
-        nickname: String
-    ) {
-
-    }
+        @ToolParam(
+            description = """
+                keyword: 유저 닉네임
+            """
+        )
+        params: PostSearchParams
+    ) = postRepository.findAllByNickname(params)
 }
